Moved DiscreteActionPolicyInterface to interfaces
